Rewrite to data scraper
scrapevaccineappt() has been rewritten to be more modular, scrape more data, and run much faster.
This function appends a dictionary to the dataframe as each Vaccine Site is processed. No more long lists of data.
This function gets the Location, Date, Address, Vaccine Type, Number of Appointments, Link to Schedule, and the Time Accessed.
The data is stored in unique csv files in the data folder for us to analyze over the period of a few weeks.
We will be able to see the trends in the data and potentially implement machine learning to predict the future availability of vaccine appointments.